requirements.txt
asgiref             == 3.7.2
Django              == 4.2.2
djangorestframework == 3.14.0
pip                 == 22.3.1
pycodestyle         == 2.10.0
pytz                == 2023.3
setuptools          == 65.5.0
sqlparse            == 0.4.4
tzdata              == 2023.3
djangorestframework-simplejwt==5.2.2